fix: adding back notifications
