+++ Client/src/lib/ui/util/CountBadge.svelte
@@ -5,12 +5,12 @@
 	export let value = 0;
 </script>
 
-<div class="absolute flex -top-1 -right-1" in:scale={{ delay: 0, duration: 250 }}>
-	<span class="absolute animate-ping w-full h-full rounded-full bg-red-600"
+<div class="absolute flex -top-1 -right-1 z-50" in:scale={{ delay: 0, duration: 250 }}>
+	<span class="absolute animate-ping w-full h-full rounded-full bg-red-800"
 		><!--ping animation--></span
 	>
 	<span
-		class="inline-flex relative rounded-full bg-red-600 text-xs px-1.5 py-[0.075rem]"
+		class="inline-flex relative rounded-full bg-red-800 text-xs px-1.5 py-[0.075rem]"
 		style="font-family: var(--header-text);"
 	>
 		<span class="relative top-[0.025rem]">{abbreviate(value)}</span>
